a bit of self-review of native.rs
